v1.3 Bug fixed: threshColor for mix calculation under m$ windows. SVG icons deleted from resources.

+++ granularmix.cpp
@@ -80,7 +80,8 @@ QImage GranularMix::blackwhiteImage() const {
 
 bool GranularMix::defThreshColor() {
 
-    threshColor = (lightThreshColor + darkThreshColor) / 2;
+    threshColor = ((unsigned long long)lightThreshColor +
+                   (unsigned long long)darkThreshColor) / 2;
 
     return true;
 }
